test: move test files
