--- packages/nuxt-swa/test/unit/runtime/types/nuxt-swa.test.ts
@@ -1,40 +1,0 @@
-// @vitest-environment node
-import { readFile } from 'node:fs/promises'
-import { resolve } from 'node:path'
-
-import { template } from 'lodash'
-import { describe, expect, it } from 'vitest'
-
-import {
-  defaults,
-  type ModuleOptions,
-  preConfiguredProviders,
-} from '~/src/config'
-
-describe('runtime/types/nuxt-swa.d.ts.template', async () => {
-  const minimumOptions: ModuleOptions = {
-    authProviders: [],
-    customRoles: [],
-    dataApi: false,
-  }
-  const defaultOptions = { ...defaults, authProviders: preConfiguredProviders }
-  const fullOptions: ModuleOptions = {
-    authProviders: ['github', 'twitter', 'line'],
-    customRoles: ['administrator'],
-    dataApi: { rest: '/rest', graphql: '/graphql' },
-  }
-  const templateString = await readFile(
-    resolve(
-      __dirname,
-      '../../../../',
-      'src/runtime/types',
-      'nuxt-swa.d.ts.template'
-    ),
-    { encoding: 'utf-8' }
-  )
-
-  it.each([defaultOptions, minimumOptions, fullOptions])(
-    '%o renders expected Type Definition',
-    options => expect(template(templateString)({ options })).toMatchSnapshot()
-  )
-})
